fix: sorting function needs to return an integer
The sorting function was returning a boolean, which is wrong.
For proper comparison it needs to return an integer.

Intrestingly enough, this issue only surfaced in the Browser
test run on Windows. There somehow the sorting function of a
two element array gets the second element first, then the
second element.